requirements.txt
asgiref==3.2.10
backports.functools-lru-cache==1.6.1
caniusepython3==7.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==2.2
django-bootstrap3==14.1.0
django-cors-headers==3.5.0
django-debug-toolbar==3.1
django-easy-pdf==0.1.1
django-extensions==3.0.9
django-rest==0.8.6
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
html5lib==1.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.50
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.4
